requirements.txt
google-cloud-storage
google-cloud-firestore
python-dotenv